+++ .github/workflows/main.yaml
@@ -36,7 +36,7 @@ jobs:
           miniconda-version: "latest"
           environment-file: conda/dev.yaml
           channels: conda-forge,nodefaults
-          activate-environment: ghreport
+          activate-environment: ghreports
           auto-update-conda: true
           conda-solver: libmamba
           python-version: "${{ matrix.python_version }}"
