+++ .github/workflows/release.yaml
@@ -25,7 +25,7 @@ jobs:
           miniconda-version: "latest"
           environment-file: conda/dev.yaml
           channels: conda-forge,nodefaults
-          activate-environment: ghreport
+          activate-environment: ghreports
           auto-update-conda: true
           conda-solver: libmamba
           python-version: "3.11"
