+++ .pre-commit-config.yaml
@@ -48,7 +48,7 @@ repos:
         name: mypy
         entry: mypy
         language: system
-        files: "ghreport/"
+        files: "ghreports/"
         pass_filenames: true
         types:
           - python
@@ -66,7 +66,7 @@ repos:
         name: vulture
         entry: vulture --min-confidence 80
         language: system
-        files: "ghreport/"
+        files: "ghreports/"
         description: Find unused Python code.
         pass_filenames: true
         types:
