+++ .releaserc.json
@@ -13,12 +13,12 @@
       {
         "replacements": [
           {
-            "files": ["src/ghreport/__init__.py"],
+            "files": ["src/ghreports/__init__.py"],
             "from": "__version__ = '.*'  # semantic-release",
             "to": "__version__ = '${nextRelease.version}'  # semantic-release",
             "results": [
               {
-                "file": "src/ghreport/__init__.py",
+                "file": "src/ghreports/__init__.py",
                 "hasChanged": true,
                 "numMatches": 1,
                 "numReplacements": 1
@@ -70,7 +70,7 @@
         "assets": [
           "pyproject.toml",
           "docs/changelog.md",
-          "src/ghreport/__init__.py"
+          "src/ghreports/__init__.py"
         ],
         "message": "chore(release): ${nextRelease.version}"
       }
