+++ README.md
@@ -3,12 +3,12 @@
 GitHub Report tool.
 
 - Free software: BSD 3 Clause
-- Documentation: https://opensciencelabs.github.io/ghreport
+- Documentation: https://opensciencelabs.github.io/ghreports
 
 ## Configuration file
 
 In order to create a configuration file, add to your project, at the root level,
-a file called .ghreport.yaml, with the following structure:
+a file called .ghreports.yaml, with the following structure:
 
 ```yaml
 name: myproject-name-slug
@@ -18,14 +18,14 @@ repos:
   - myorg-1/myproject1
 authors:
   - gh-username-1: GitHub Username 1
-output-dir: "/tmp/ghreport"
+output-dir: "/tmp/ghreports"
 ```
 
-## How to run the ghreport
+## How to run the ghreports
 
 ```bash
 
-ghreport --start-date 2025-07-01 --end-date 2025-07-31 --config-file tests/.ghreport.yaml
+ghreports --start-date 2025-07-01 --end-date 2025-07-31 --config-file tests/.ghreports.yaml
 
 ```
 
@@ -33,6 +33,6 @@ You can also specify the token in the command line as an argument:
 
 ```bash
 
-ghreport --start-date 2025-07-01 --end-date 2025-07-31 --gh-token blabla --config-file tests/.ghreport.yaml
+ghreports --start-date 2025-07-01 --end-date 2025-07-31 --gh-token blabla --config-file tests/.ghreports.yaml
 
 ```
