+++ docs/installation.md
@@ -5,7 +5,7 @@
 To install GHReport, run the following command in your terminal:
 
 ```bash
-$ pip install ghreport
+$ pip install ghreports
 ```
 
 These installation methods will always provide you with the most recent stable
@@ -13,20 +13,20 @@ release.
 
 ## From sources
 
-The sources for ghreport can be downloaded from the
-[Github repo](https://github.com/osl-incubator/ghreport.git).
+The sources for ghreports can be downloaded from the
+[Github repo](https://github.com/osl-incubator/ghreports.git).
 
 You can either clone the public repository:
 
 ```bash
-$ git clone https://github.com/osl-incubator/ghreport.git
+$ git clone https://github.com/osl-incubator/ghreports.git
 ```
 
 Or download the
-[tarball](https://github.com/osl-incubator/ghreport.git/tarball/main):
+[tarball](https://github.com/osl-incubator/ghreports.git/tarball/main):
 
 ```bash
-$ curl -OJL https://github.com/osl-incubator/ghreport.git/tarball/main
+$ curl -OJL https://github.com/osl-incubator/ghreports.git/tarball/main
 ```
 
 Once you have a copy of the source, change to the project root directory and
