+++ src/ghreports/templates/search.graphql
