+++ tests/test_ghreport.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
-from ghreport import GHReport
-from ghreport.config import ArgsCLI
+from ghreports import GHReport
+from ghreports.config import ArgsCLI
 
 
 def test_ghreport():
@@ -10,7 +10,7 @@ def test_ghreport():
             **{
                 'start_date': '2023-07-01',
                 'end_date': '2023-07-31',
-                'config_file': str(Path(__file__).parent / '.ghreport.yaml'),
+                'config_file': str(Path(__file__).parent / '.ghreports.yaml'),
             }
         )
     )
